permission for inventory material

+++ src/app/views/businesses/stores/inventory-materials/inventory-material-list/inventory-material-list.component.ts
@@ -8,6 +8,7 @@ import { NotifyService } from 'src/app/shared/services/notify.service';
 import { Pagination, PaginatedResult } from 'src/app/shared/models/pagination.model';
 import { PagingParams } from 'src/app/shared/params/paging.param';
 import { InventoryMaterial } from 'src/app/shared/models/inventory-material.model';
+import { RoleService } from 'src/app/shared/services/role.service';
 // import {
 //   LiquidationMaterialViewDetailModalComponent
 // } from '../modals/liquidation-material-view-detail-modal/liquidation-material-view-detail-modal.component';
@@ -39,13 +40,21 @@ export class InventoryMaterialListComponent implements OnInit {
   @HostListener('window:keydown', ['$event'])
   onKeyPress($event: KeyboardEvent) {
     if (($event.ctrlKey || $event.metaKey) && ($event.keyCode === 73 || $event.keyCode === 105)) {
-      this.router.navigate(['/nghiep-vu/kho/phieu-kiem-ke-kho/tao-phieu-kiem-ke']);
+      this.roleService.checkPermission('PHIEU_KIEM_KE_KHO', 'Create')
+        .subscribe((response: boolean) => {
+          if (response) {
+            this.router.navigate(['/nghiep-vu/kho/phieu-kiem-ke-kho/tao-phieu-kiem-ke']);
+          } else {
+            this.notify.warning('Bạn không có quyền');
+          }
+        });
     }
   }
 
   constructor(private route: ActivatedRoute,
     private router: Router,
     private modalService: NzModalService,
+    private roleService: RoleService,
     private inventoryMaterialService: InventoryMaterialService,
     private notify: NotifyService) { }
 
@@ -73,13 +82,32 @@ export class InventoryMaterialListComponent implements OnInit {
         this.loading = false;
         this.pagination = res.pagination;
         this.dataSet = res.result;
-      }, error => {
-        this.notify.error('Có lỗi xảy ra');
-        console.log('error getAllPagingInventoryMaterial');
-      }, () => {
+
         if (this.dataSet.length === 0 && this.pagination.currentPage !== 1) {
           this.pagination.currentPage -= 1;
           this.loadData();
+        }
+      });
+  }
+
+  addNew() {
+    this.roleService.checkPermission('PHIEU_KIEM_KE_KHO', 'Create')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.router.navigate(['/nghiep-vu/kho/phieu-kiem-ke-kho/tao-phieu-kiem-ke']);
+        } else {
+          this.notify.warning('Bạn không có quyền');
+        }
+      });
+  }
+
+  update(id: number) {
+    this.roleService.checkPermission('PHIEU_KIEM_KE_KHO', 'Update')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.router.navigate(['/nghiep-vu/kho/phieu-kiem-ke-kho/sua-phieu-kiem-ke', id]);
+        } else {
+          this.notify.warning('Bạn không có quyền');
         }
       });
   }
@@ -107,16 +135,23 @@ export class InventoryMaterialListComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
-      this.inventoryMaterialService.delete(id).subscribe((res: boolean) => {
-        if (res) {
-          this.loadData();
-          this.notify.success('Xóa thành công');
+    this.roleService.checkPermission('PHIEU_KIEM_KE_KHO', 'Delete')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
+            this.inventoryMaterialService.delete(id).subscribe((res: boolean) => {
+              if (res) {
+                this.loadData();
+                this.notify.success('Xóa thành công');
+              } else {
+                this.notify.warning('Không được xóa vì đã xuất hoặc thanh lý');
+              }
+            });
+          });
         } else {
-          this.notify.warning('Không được xóa vì đã xuất hoặc thanh lý');
+          this.notify.warning('Bạn không có quyền');
         }
       });
-    });
   }
 
   search() {
